refactor/lm_experiments_tools (#12)
* refactor: move trainer and utils to lm_exp_tools package

* docs: update readme

* docs: readme updates
