+++ requirements-lm-tools.txt
@@ -1,0 +1,5 @@
+transformers>=4.17.0
+torch>=1.7.1
+numpy>=1.19.4
+tqdm>=4.41.1
+horovod>=0.24.2
